+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2.9.9
+- Updated: Item Adjustment Incoming - August 25
+- Updated: Item Adjustment Incoming - 26 August
+
 ## 2.9.8
 - Updated: Item Adjustment Incoming - August 18
 
@@ -835,6 +839,7 @@
 - Added external link copy and guild chat sharing for selected posts.
 - Added login toasts for recent posts.
 - Added class section shortcuts for class hotfix posts.
+